backend/requirements_db.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
python-multipart==0.0.6
python-dotenv==1.0.0
PyJWT==2.10.1
httpx==0.25.2
requests==2.31.0
APScheduler==3.10.4
redis==5.0.1
aioredis==2.0.1
